pyproject.toml
[project]
name = "super-mario"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "glfw>=2.10.0",
    "pillow>=12.2.0",
    "pyopengl>=3.1.10",
]
